+++ packages/terra-data-grid/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+* Removed
+  * Removed the `boundingRef` prop in favor of using the table's container ref for calculating the offset. is no longer needed for the section header's sticky behavior. This change is still passive in nature as it doesn't change any functionality for consumers.
+
+
 ## 1.20.0 - (March 14, 2024)
 
 * Added